Restructure viewset test case
